+++ hwloc/topology-solaris.c
@@ -417,11 +417,11 @@ hwloc_look_lgrp(struct hwloc_topology *topology)
     browse(topology, cookie, root, glob_lgrps, &curlgrp);
 #if HAVE_DECL_LGRP_LATENCY_COOKIE
     if (nlgrps > 1) {
-      float *distances = calloc(curlgrp*curlgrp, sizeof(float));
+      uint64_t *distances = calloc(curlgrp*curlgrp, sizeof(uint64_t));
       unsigned i, j;
       for (i = 0; i < curlgrp; i++) {
 	for (j = 0; j < curlgrp; j++)
-          distances[i*curlgrp+j] = (float) lgrp_latency_cookie(cookie, glob_lgrps[i]->os_index, glob_lgrps[j]->os_index, LGRP_LAT_CPU_TO_MEM);
+          distances[i*curlgrp+j] = (uint64_t) lgrp_latency_cookie(cookie, glob_lgrps[i]->os_index, glob_lgrps[j]->os_index, LGRP_LAT_CPU_TO_MEM);
       }
       hwloc_internal_distances_add(topology, curlgrp, glob_lgrps, distances,
 				   HWLOC_DISTANCES_KIND_FROM_OS|HWLOC_DISTANCES_KIND_MEANS_LATENCY,
